+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Unreleased
 ==========
 
 * Update AWS provider required version to ~> 3.0
+* Checkov baseline static analysis.  All issues flagged by Checkov have *not*
+  yet been resolved.
 
 1.1.0
 =====