+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 * Checkov baseline static analysis.  All issues flagged by Checkov have *not*
   yet been resolved.
 * Status badges for validation and static analysis
+* Improved naming of autoscaling group and launch configuration
 
 1.1.0
 =====